Code split guidelines followed
